Move XCScheme to Xcode Kit project
