+++ subprojects/xcode-ide-kit/xcode-ide-kit.gradle
@@ -5,6 +5,7 @@ plugins {
 }
 
 dependencies {
+	implementation 'com.fasterxml.jackson.dataformat:jackson-dataformat-xml:2.9.8'
 	implementation "com.google.guava:guava:${guavaVersion}"
 	implementation "org.antlr:antlr4:${antlrVersion}"
 	antlr "org.antlr:antlr4:${antlrVersion}"
